# api-docs.json
{"openapi":"3.0.1","info":{"title":"Drive API","version":"v1.0"},"servers":[{"url":"https://localhost:8080"}],"paths":{"/api/drive/files/{fileId}/permission/revoke/{permissionId}":{"post":{"tags":["permission-controller"],"operationId":"revokePermission","parameters":[{"name":"permissionId","in":"path","required":true,"schema":{"type":"string"}},{"name":"fileId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/drive/files/{fileId}/permission/add":{"post":{"tags":["permission-controller"],"operationId":"addPermission","parameters":[{"name":"fileId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PermissionDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/drive/files/upload":{"post":{"tags":["file-controller"],"operationId":"uploadFiles","parameters":[{"name":"files","in":"query","required":true,"schema":{"type":"array","items":{"type":"string","format":"binary"}}},{"name":"targetFolderId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/api/drive/files/untrash":{"post":{"tags":["trash-controller"],"description":"Remove files from trash","operationId":"untrashFiles","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/drive/files/trash":{"post":{"tags":["trash-controller"],"description":"Add files to trash","operationId":"trashFiles","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["trash-controller"],"description":"Delete all files from trash","operationId":"emptyTrash","responses":{"200":{"description":"OK"}}}},"/api/drive/files/delete":{"post":{"tags":["file-controller"],"description":"Delete files completely","operationId":"deleteFiles","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/drive/files/create/folder":{"post":{"tags":["file-controller"],"operationId":"createFolder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FolderDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/drive/sign-in":{"get":{"tags":["auth-controller"],"operationId":"googleSignIn","responses":{"200":{"description":"OK"}}}},"/api/drive/oauth":{"get":{"tags":["auth-controller"],"operationId":"saveAuthorizationCode","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/drive/files":{"get":{"tags":["file-controller"],"operationId":"listFiles","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GoogleFileDto"}}}}}}}},"/api/drive/files/get/{fileId}":{"get":{"tags":["file-controller"],"operationId":"getFile","parameters":[{"name":"fileId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GoogleFileShortDto"}}}}}}},"/api/drive/files/download":{"get":{"tags":["file-controller"],"operationId":"downloadFiles","parameters":[{"name":"download","in":"query","required":true,"schema":{"type":"object","additionalProperties":{"type":"string"}},"example":{"fileId1":"absolutePath1","fileId2":"absolutePath2"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"PermissionDto":{"type":"object","properties":{"role":{"type":"string","enum":["organizer","fileOrganizer","writer","commenter","reader"]},"type":{"type":"string","enum":["user","group","domain","anyone"]},"emailAddressOrDomain":{"type":"string"}}},"FolderDto":{"type":"object","properties":{"targetFolderId":{"type":"string"},"folderName":{"type":"string"}}},"ClassInfo":{"type":"object","properties":{"ignoreCase":{"type":"boolean"},"names":{"type":"array","items":{"type":"string"}},"enum":{"type":"boolean"},"fieldInfos":{"type":"array","items":{"$ref":"#/components/schemas/FieldInfo"}}}},"DateTime":{"type":"object","properties":{"value":{"type":"integer","format":"int64"},"dateOnly":{"type":"boolean"},"timeZoneShift":{"type":"integer","format":"int32"}}},"FieldInfo":{"type":"object","properties":{"field":{"type":"object","properties":{"name":{"type":"string"},"modifiers":{"type":"integer","format":"int32"},"synthetic":{"type":"boolean"},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"accessible":{"type":"boolean","deprecated":true},"genericType":{"type":"object","properties":{"typeName":{"type":"string"}}},"enumConstant":{"type":"boolean"},"annotatedType":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}},"annotations":{"type":"array","items":{"type":"object"}}}},"name":{"type":"string"},"primitive":{"type":"boolean"},"genericType":{"type":"object","properties":{"typeName":{"type":"string"}}},"final":{"type":"boolean"},"classInfo":{"$ref":"#/components/schemas/ClassInfo"},"tersMethodForField":{"type":"array","writeOnly":true,"items":{"type":"object","properties":{"name":{"type":"string"},"modifiers":{"type":"integer","format":"int32"},"typeParameters":{"type":"array","items":{"type":"object","properties":{"name":{"type":"string"},"bounds":{"type":"array","items":{"type":"object","properties":{"typeName":{"type":"string"}}}},"annotatedBounds":{"type":"array","items":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}}},"typeName":{"type":"string"},"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}}}}},"synthetic":{"type":"boolean"},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"accessible":{"type":"boolean","deprecated":true},"varArgs":{"type":"boolean"},"parameterCount":{"type":"integer","format":"int32"},"parameterAnnotations":{"type":"array","items":{"type":"array","items":{"type":"object"}}},"genericParameterTypes":{"type":"array","items":{"type":"object","properties":{"typeName":{"type":"string"}}}},"genericExceptionTypes":{"type":"array","items":{"type":"object","properties":{"typeName":{"type":"string"}}}},"default":{"type":"boolean"},"genericReturnType":{"type":"object","properties":{"typeName":{"type":"string"}}},"bridge":{"type":"boolean"},"defaultValue":{"type":"object"},"annotatedReturnType":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}},"annotatedParameterTypes":{"type":"array","items":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}}},"parameters":{"type":"array","items":{"type":"object","properties":{"name":{"type":"string"},"modifiers":{"type":"integer","format":"int32"},"synthetic":{"type":"boolean"},"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"annotatedType":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}},"parameterizedType":{"type":"object","properties":{"typeName":{"type":"string"}}},"varArgs":{"type":"boolean"},"namePresent":{"type":"boolean"},"declaringExecutable":{"type":"object","properties":{"name":{"type":"string"},"modifiers":{"type":"integer","format":"int32"},"typeParameters":{"type":"array","items":{"type":"object","properties":{"name":{"type":"string"},"bounds":{"type":"array","items":{"type":"object","properties":{"typeName":{"type":"string"}}}},"genericDeclaration":{"type":"object"},"annotatedBounds":{"type":"array","items":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}}},"typeName":{"type":"string"},"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}}}}},"synthetic":{"type":"boolean"},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"varArgs":{"type":"boolean"},"annotatedParameterTypes":{"type":"array","items":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}}},"parameterCount":{"type":"integer","format":"int32"},"parameterAnnotations":{"type":"array","items":{"type":"array","items":{"type":"object"}}},"genericParameterTypes":{"type":"array","items":{"type":"object","properties":{"typeName":{"type":"string"}}}},"genericExceptionTypes":{"type":"array","items":{"type":"object","properties":{"typeName":{"type":"string"}}}},"annotatedReturnType":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}},"annotatedReceiverType":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}},"annotatedExceptionTypes":{"type":"array","items":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}}},"annotations":{"type":"array","items":{"type":"object"}},"accessible":{"type":"boolean","deprecated":true}}},"implicit":{"type":"boolean"}}}},"annotatedReceiverType":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}},"annotatedExceptionTypes":{"type":"array","items":{"type":"object","properties":{"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"type":{"type":"object","properties":{"typeName":{"type":"string"}}}}}},"annotations":{"type":"array","items":{"type":"object"}}}}}}},"GoogleFileDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"mimeType":{"type":"string"},"createdTime":{"type":"string","format":"date-time"},"modifiedTime":{"type":"string","format":"date-time"},"permissions":{"type":"array","items":{"$ref":"#/components/schemas/Permission"}},"trashed":{"type":"boolean"},"size":{"type":"integer","format":"int64"},"parents":{"type":"array","items":{"type":"string"}}}},"JsonFactory":{"type":"object"},"Permission":{"type":"object","properties":{"classInfo":{"$ref":"#/components/schemas/ClassInfo"},"allowFileDiscovery":{"type":"boolean"},"deleted":{"type":"boolean"},"displayName":{"type":"string"},"domain":{"type":"string"},"emailAddress":{"type":"string"},"expirationTime":{"$ref":"#/components/schemas/DateTime"},"id":{"type":"string"},"kind":{"type":"string"},"pendingOwner":{"type":"boolean"},"permissionDetails":{"type":"array","items":{"$ref":"#/components/schemas/PermissionDetails"}},"photoLink":{"type":"string"},"role":{"type":"string"},"teamDrivePermissionDetails":{"type":"array","items":{"$ref":"#/components/schemas/TeamDrivePermissionDetails"}},"type":{"type":"string"},"view":{"type":"string"},"factory":{"$ref":"#/components/schemas/JsonFactory"},"unknownKeys":{"type":"object","additionalProperties":{"type":"object"}},"empty":{"type":"boolean"}},"additionalProperties":{"type":"object"}},"PermissionDetails":{"type":"object","properties":{"classInfo":{"$ref":"#/components/schemas/ClassInfo"},"inherited":{"type":"boolean"},"inheritedFrom":{"type":"string"},"permissionType":{"type":"string"},"role":{"type":"string"},"factory":{"$ref":"#/components/schemas/JsonFactory"},"unknownKeys":{"type":"object","additionalProperties":{"type":"object"}},"empty":{"type":"boolean"}},"additionalProperties":{"type":"object"}},"TeamDrivePermissionDetails":{"type":"object","properties":{"classInfo":{"$ref":"#/components/schemas/ClassInfo"},"inherited":{"type":"boolean"},"inheritedFrom":{"type":"string"},"role":{"type":"string"},"teamDrivePermissionType":{"type":"string"},"factory":{"$ref":"#/components/schemas/JsonFactory"},"unknownKeys":{"type":"object","additionalProperties":{"type":"object"}},"empty":{"type":"boolean"}},"additionalProperties":{"type":"object"}},"GoogleFileShortDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"mimeType":{"type":"string"}}}},"securitySchemes":{"security_auth":{"type":"oauth2","flows":{"authorizationCode":{"authorizationUrl":"http://localhost:8080/api/drive/sign-in","tokenUrl":"http://localhost:8080/api/drive/oauth","scopes":{}}}}}}}